+++ barretenberg/cpp/src/barretenberg/commitment_schemes/ipa/ipa.test.cpp
@@ -1,13 +1,7 @@
 
-#include "../gemini/gemini.hpp"
-#include "../shplonk/shplemini.hpp"
-#include "../shplonk/shplonk.hpp"
 #include "./mock_transcript.hpp"
 #include "barretenberg/commitment_schemes/commitment_key.test.hpp"
-#include "barretenberg/common/mem.hpp"
-#include "barretenberg/ecc/curves/bn254/fq12.hpp"
-#include "barretenberg/ecc/curves/types.hpp"
-#include "barretenberg/polynomials/polynomial_arithmetic.hpp"
+#include "barretenberg/commitment_schemes/shplonk/shplemini.hpp"
 using namespace bb;
 
 namespace {
@@ -55,7 +49,7 @@ TEST_F(IPATest, OpenZeroPolynomial)
     constexpr size_t n = 4;
     Polynomial poly(n);
     // Commit to a zero polynomial
-    GroupElement commitment = this->commit(poly);
+    Commitment commitment = this->commit(poly);
     EXPECT_TRUE(commitment.is_point_at_infinity());
 
     auto [x, eval] = this->random_eval(poly);
@@ -247,8 +241,8 @@ TEST_F(IPATest, GeminiShplonkIPAWithShift)
     auto poly1 = Polynomial::random(n);
     auto poly2 = Polynomial::random(n, /*shiftable*/ 1);
 
-    GroupElement commitment1 = this->commit(poly1);
-    GroupElement commitment2 = this->commit(poly2);
+    Commitment commitment1 = this->commit(poly1);
+    Commitment commitment2 = this->commit(poly2);
 
     auto eval1 = poly1.evaluate_mle(mle_opening_point);
     auto eval2 = poly2.evaluate_mle(mle_opening_point);
@@ -272,7 +266,8 @@ TEST_F(IPATest, GeminiShplonkIPAWithShift)
     auto verifier_transcript = NativeTranscript::verifier_init_empty(prover_transcript);
 
     auto gemini_verifier_claim = GeminiVerifier::reduce_verification(mle_opening_point,
-                                                                     multilinear_evaluations,
+                                                                     RefArray{ eval1, eval2 },
+                                                                     RefArray{ eval2_shift },
                                                                      RefArray{ commitment1, commitment2 },
                                                                      RefArray{ commitment2 },
                                                                      verifier_transcript);