+++ barretenberg/cpp/src/barretenberg/relations/translator_vm/translator_permutation_relation.cpp
@@ -2,5 +2,5 @@
 #include "barretenberg/translator_vm/translator_flavor.hpp"
 namespace bb {
 template class TranslatorPermutationRelationImpl<fr>;
-DEFINE_SUMCHECK_RELATION_CLASS(TranslatorPermutationRelationImpl, TranslatorFlavor);
+DEFINE_ZK_SUMCHECK_RELATION_CLASS(TranslatorPermutationRelationImpl, TranslatorFlavor);
 } // namespace bb