+++ l1-contracts/src/core/interfaces/IRollup.sol
@@ -105,7 +105,7 @@ interface IRollup {
   function getProvenBlockNumber() external view returns (uint256);
   function getPendingBlockNumber() external view returns (uint256);
   function getEpochToProve() external view returns (Epoch);
-  function nextEpochToClaim() external view returns (Epoch);
+  function getClaimableEpoch() external view returns (Epoch);
   function getEpochForBlock(uint256 blockNumber) external view returns (Epoch);
   function validateEpochProofRightClaim(EpochProofQuoteLib.SignedEpochProofQuote calldata _quote)
     external