+++ yarn-project/bb-prover/src/bb/execute.ts
@@ -504,7 +504,6 @@ export async function generateAvmProof(
   }
 
   // Paths for the inputs
-  const bytecodePath = join(workingDirectory, AVM_BYTECODE_FILENAME);
   const calldataPath = join(workingDirectory, AVM_CALLDATA_FILENAME);
   const publicInputsPath = join(workingDirectory, AVM_PUBLIC_INPUTS_FILENAME);
   const avmHintsPath = join(workingDirectory, AVM_HINTS_FILENAME);
@@ -525,10 +524,6 @@ export async function generateAvmProof(
 
   try {
     // Write the inputs to the working directory.
-    await fs.writeFile(bytecodePath, input.bytecode);
-    if (!filePresent(bytecodePath)) {
-      return { status: BB_RESULT.FAILURE, reason: `Could not write bytecode at ${bytecodePath}` };
-    }
     await fs.writeFile(
       calldataPath,
       input.calldata.map(fr => fr.toBuffer()),
@@ -553,8 +548,6 @@ export async function generateAvmProof(
     }
 
     const args = [
-      '--avm-bytecode',
-      bytecodePath,
       '--avm-calldata',
       calldataPath,
       '--avm-public-inputs',