+++ yarn-project/end-to-end/src/e2e_event_logs.test.ts
@@ -6,7 +6,6 @@ import {
   L1EventPayload,
   type PXE,
 } from '@aztec/aztec.js';
-import { deriveMasterIncomingViewingSecretKey } from '@aztec/circuits.js';
 import { EventSelector } from '@aztec/foundation/abi';
 import { makeTuple } from '@aztec/foundation/array';
 import { type Tuple } from '@aztec/foundation/serialize';
@@ -55,10 +54,7 @@ describe('Logs', () => {
       const encryptedLogs = txEffect!.encryptedLogs.unrollLogs();
       expect(encryptedLogs.length).toBe(3);
 
-      const decryptedEvent0 = L1EventPayload.decryptAsIncoming(
-        encryptedLogs[0],
-        deriveMasterIncomingViewingSecretKey(wallets[0].getSecretKey()),
-      )!;
+      const decryptedEvent0 = L1EventPayload.decryptAsIncoming(encryptedLogs[0], wallets[0].getEncryptionSecret())!;
 
       expect(decryptedEvent0.contractAddress).toStrictEqual(testLogContract.address);
       expect(decryptedEvent0.randomness).toStrictEqual(randomness[0]);
@@ -75,10 +71,7 @@ describe('Logs', () => {
       const badEvent0 = TestLogContract.events.ExampleEvent1.decode(decryptedEvent0);
       expect(badEvent0).toBe(undefined);
 
-      const decryptedEvent1 = L1EventPayload.decryptAsIncoming(
-        encryptedLogs[2],
-        deriveMasterIncomingViewingSecretKey(wallets[0].getSecretKey()),
-      )!;
+      const decryptedEvent1 = L1EventPayload.decryptAsIncoming(encryptedLogs[2], wallets[0].getEncryptionSecret())!;
 
       expect(decryptedEvent1.contractAddress).toStrictEqual(testLogContract.address);
       expect(decryptedEvent1.randomness).toStrictEqual(randomness[1]);