+++ yarn-project/sequencer-client/src/sequencer/sequencer.ts
@@ -40,7 +40,7 @@ import { inspect } from 'util';
 import { type BlockBuilderFactory } from '../block_builder/index.js';
 import { type GlobalVariableBuilder } from '../global_variable_builder/global_builder.js';
 import { type L1Publisher } from '../publisher/l1-publisher.js';
-import { prettyLogViemError } from '../publisher/utils.js';
+import { prettyLogViemErrorMsg } from '../publisher/utils.js';
 import { type TxValidatorFactory } from '../tx_validator/tx_validator_factory.js';
 import { type SequencerConfig } from './config.js';
 import { SequencerMetrics } from './metrics.js';
@@ -137,6 +137,10 @@ export class Sequencer {
     if (config.allowedInTeardown) {
       this.allowedInTeardown = config.allowedInTeardown;
     }
+    if (config.gerousiaPayload) {
+      this.publisher.setPayload(config.gerousiaPayload);
+    }
+
     // TODO: Just read everything from the config object as needed instead of copying everything into local vars.
     this.config = config;
   }
@@ -223,6 +227,15 @@ export class Sequencer {
       return;
     }
 
+    const newGlobalVariables = await this.globalsBuilder.buildGlobalVariables(
+      new Fr(newBlockNumber),
+      this._coinbase,
+      this._feeRecipient,
+      slot,
+    );
+
+    void this.publisher.castVote(slot, newGlobalVariables.timestamp.toBigInt());
+
     if (!this.shouldProposeBlock(historicalHeader, {})) {
       return;
     }
@@ -236,13 +249,6 @@ export class Sequencer {
       return;
     }
     this.log.debug(`Retrieved ${pendingTxs.length} txs from P2P pool`);
-
-    const newGlobalVariables = await this.globalsBuilder.buildGlobalVariables(
-      new Fr(newBlockNumber),
-      this._coinbase,
-      this._feeRecipient,
-      slot,
-    );
 
     // If I created a "partial" header here that should make our job much easier.
     const proposalHeader = new Header(
@@ -309,8 +315,10 @@ export class Sequencer {
       this.log.debug(`Can propose block ${proposalBlockNumber} at slot ${slot}`);
       return slot;
     } catch (err) {
-      this.log.verbose(`Rejected from being able to propose at next block with ${tipArchive}`);
-      prettyLogViemError(err, this.log);
+      const msg = prettyLogViemErrorMsg(err);
+      this.log.verbose(
+        `Rejected from being able to propose at next block with ${tipArchive.toString('hex')}: ${msg ? `${msg}` : ''}`,
+      );
       throw err;
     }
   }
@@ -548,20 +556,12 @@ export class Sequencer {
   protected async createProofClaimForPreviousEpoch(slotNumber: bigint): Promise<EpochProofQuote | undefined> {
     try {
       // Find out which epoch we are currently in
-      const epochForBlock = await this.publisher.getEpochForSlotNumber(slotNumber);
-      if (epochForBlock < 1n) {
-        // It's the 0th epoch, nothing to be proven yet
-        this.log.verbose(`First epoch has no claim`);
+      const epochToProve = await this.publisher.getClaimableEpoch();
+      if (epochToProve === undefined) {
+        this.log.verbose(`No epoch to prove`);
         return undefined;
       }
-      const epochToProve = epochForBlock - 1n;
-      // Find out the next epoch that can be claimed
-      const canClaim = await this.publisher.nextEpochToClaim();
-      if (canClaim != epochToProve) {
-        // It's not the one we are looking to claim
-        this.log.verbose(`Unable to claim previous epoch (${canClaim} != ${epochToProve})`);
-        return undefined;
-      }
+
       // Get quotes for the epoch to be proven
       const quotes = await this.p2pClient.getEpochProofQuotes(epochToProve);
       this.log.verbose(`Retrieved ${quotes.length} quotes, slot: ${slotNumber}, epoch to prove: ${epochToProve}`);