+++ yarn-project/simulator/src/public/side_effect_trace.test.ts
@@ -40,7 +40,7 @@ describe('Side Effect Trace', () => {
   const recipient = Fr.random();
   const content = Fr.random();
   const log = [Fr.random(), Fr.random(), Fr.random()];
-  const contractInstance = SerializableContractInstance.empty().withAddress(new Fr(42));
+  const contractInstance = SerializableContractInstance.default().withAddress(new Fr(42));
 
   const startGasLeft = Gas.fromFields([new Fr(randomInt(10000)), new Fr(randomInt(10000))]);
   const endGasLeft = Gas.fromFields([new Fr(randomInt(10000)), new Fr(randomInt(10000))]);
@@ -243,7 +243,7 @@ describe('Side Effect Trace', () => {
       {
         // hint omits "version" and has "exists" as an Fr
         ...instanceWithoutVersion,
-        exists: new Fr(instance.exists),
+        exists: instance.exists,
       },
     ]);
   });